Project_2/requirements.txt
streamlit
pandas
openai
python-dotenv
python-dotenv